+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - 流程表达式 follow后端更新
 - websocket功能(默认关闭)
+- useVbenForm 增加 TimeRangePicker(时间区间选择) 组件
 
 **REFACTOR**